content/canvas/test/test_2d.imageData.object.undefined.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.object.undefined</title>
<!-- Testing: ImageData.data converts undefined to 0 -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var imgdata = ctx.getImageData(0, 0, 10, 10);
imgdata.data[0] = 100;
imgdata.data[0] = undefined;
todo(imgdata.data[0] === 0, "imgdata.data[\""+(0)+"\"] === 0");

SimpleTest.finish();

});
</script>
